Don't use require for metadata.json
An error was detected while testing with a path that is not explicitly
absolute or relative, which results into requiring from node_modules.
e.g. require("path/to/metadata.json") would try to find
"to/metadata.json" within the "path" package.

This might be solve by using path.resolve on the path passed to require.
But this still leaves a problem open, which is a secuirity risk of
requiring a file that is dynamically downloaded. If it doesn't contain
JSON but JavaScript it would be executed directly.

Therefore, using fs.readFile + JSON.parse is a much better approach.