[INTERNAL] npm translator: Improve detection of collection projects
A collection must have an object property "collection" with a nested
object "modules" in its package.json.
Other types or structures should be ignored and the project should not
be handled as collection.

This might happen when other projects define a "collection" property
that is meant to be used for a different purpose
(e.g. https://stenciljs.com/docs/publishing#package-json).