[FIX] ui5Framework.Installer: Ensure target directory does not exist before rename
Renaming a directory might fail with an EPERM error on Windows if the
target already exists. Therefore we stop creating the target directory
before the rename (this was probably the main problem) and also ensure
that it does not exist for whatever other reasons.

Related: https://github.com/SAP/ui5-project/pull/389
Resolves: https://github.com/SAP/ui5-tooling/issues/487