[INTERNAL] ProjectGraph: Remove @ui5/builder dependency
With this change, a 'taskRepository' instance is required to be passed
to the #build function of a project graph (dependency injection).
This removes the requirement of having a direct @ui5/builder dependency
in @ui5/project

In the future, we might add checks to ensure version compatibility with
the injected module.

Depends on https://github.com/SAP/ui5-builder/pull/842