[INTERNAL] Workspace: Match framework projects based on name instead of module ID
This should make the use of workspaces easier. Framework projects are
matched based on the name in the ui5.yaml instead of the package.json
name (= module ID). This allows for example framework projects to be
contained in packages that are named differently from the local copy.
E.g. "@openui5/sap.m-prebuilt" while the local copy does not have the
"-prebuilt" suffix.

Note however, that the NodePackageDependencies-provider still matches
nodes using the module ID.

The basic reason for this difference is that framework dependencies are
typically declared using the project name, while npm dependencies are
declared using the module ID.

However, we could also enhance projectGraphBuilder in the future to
additionally check whether a project can be resolved with a workspace
once the project has been created and its name is known. This could
catch those cases where the node provider could not find a workspace
resolution for a given module ID.