[INTERNAL] TaskRunner: Custom Tasks receive 'dependencies' reader in most cases
Always provide a 'dependencies'-reader to custom tasks, unless they
define Specification Version >=3.0 and do not provide a
determineRequiredDependencies callback.

The latter being an indicator that either dependencies are really not
required by the task or that the concept of the callback is not yet
known or applied incorrectly. By not providing any dependencies reader
in this case, the task will face an exception if access to the reader is
still attempted, for example via 'dependencies.byGlob()' (=> 'byGlob of
undefined').

At the same time, tasks implementing the callback but not requesting any
dependencies will receive an empty 'dependencies'-reader. This can ease
the implementation efforts in the task.

Resolves https://github.com/SAP/ui5-tooling/issues/732