[INTERNAL] Allow framework projects use of framework resolver
If the current root project is a framework project (identified by the
scope in the package.json name), allow processing of it's framework
dependencies by the ui5Framework helper module. Other framework projects
in the graph are still ignored.

This allows framework developers to retrieve all required dependencies
in a given version. The version is either defined in the ui5.yaml (not
recommended) or via the --framework-version CLI parameter.

If no version is specified, all required framework dependencies must be
part of the presented graph or an error is thrown.