[FIX] graph: Always resolve rootConfigPath to CWD
This should restore the v2 behavior where referencing a ui5.yaml file
via the '--config' CLI parameter can always be done with a path relative
to the directory the CLI has been invoked in.

v3 used to pass the relative path to graph/Module which would resolve it
against the module's root path (typically the location of the
package.json) using @ui5/fs.

This can be unexpected, for example if the CLI is invoked in a
subdirectory of the module. Also, referencing configuration files
outside of the module (e.g. "../ui5.yaml") was no longer possible.