[FIX] projectGraphBuilder: Apply extensions of the same module only once
The same module will be processed every time another module depends on
it. This change ensures that any extensions contained in the module will
only be applied once and only if the dependency is non-optional or if
the parent project is the root project.

Fixes https://github.com/SAP/ui5-tooling/issues/777