[FIX] projectGraphBuilder: Add module cache invalidation
If a visited module did not resolve to any specification (project or
extensions), try recreating the module when visiting it again from a
different parent.

Resolves https://github.com/SAP/ui5-tooling/issues/807

This is an alternative to https://github.com/SAP/ui5-project/pull/611

This change might have a greater impact on performance since er might
recreate modules that are not relevant to UI5 Tooling more than once if
they are listed multiple times in the dependency tree.
Before this change, such modules where only visited once.