[FIX] ui5Framework: Treat 'optional' dependencies of root project as non-optional
This is a v3 regression. Optional dependencies of the root project
should always be enforced (just like development dependencies).

Instead, ui5Framework would correctly install them, but then declared the
dependency in the graph as 'optional'. Which meant the dependency would
only be enforced if another dependency would define a non-optional
dependency to it.

BCP: 2380060321