[INTERNAL] Sapui5MavenSnapshotResolver: Store npm-dist.zip contents under different name from default JAR
This is to ensure compatibility with old UI5 Tooling versions that are
used in parallel. An old version might download the default JAR for
1.117.0-SNAPSHOT and store it under "<library name>-prebuilt".

While a new UI5 Tooling version would expect the content of the
npm-dist.zip artifact under that name.

Therefore, we use a new name for storing the new artifact.