[INTERNAL] Improve compatibility handling
with installed content from previous ui5-project versions.

The prebuilt package previously didn't contain a package.json.
Now we only consider a package to be installed when it has a package.json
This causes old jar based packages that already exist to be removed and
replaced by the new npm-dist.zip contents.