[POC] ProjectBuilder: Build projects in parallel if possible
Based on the required dependencies information provided by the tasks,
determine which projects need to be build in sequence and which can be
build in parallel.

This can lead to a significantly faster build in some cases, depending
on the number of dependencies and which tasks need to be executed.