[FIX] Allow usage of after/before task assignment of generateBundle
When migrating an project from V2 to V3 the after/before task
assignments of custom tasks often has to be adjusted because the task
`uglify` got removed and new `minify` task is executed earlier as the
`uglify` task.

In the migration guide we have written that we can not make any
assumptions about the task which should be used as the new "last" task
executed by the standard UI5 Tooling build tasks.
See https://sap.github.io/ui5-tooling/stable/updates/migrate-v3/#removal-of-standard-tasks-and-processors.

In the above mentioned documentation we have a list with all tasks and
the execution order.
Currently, it works to assign a custom task after/before a task which
is not executed at all.
So you can assign a custom task to `afterTask: generateVersionInfo`
although the task `generateVersionInfo` is not executed.

This works for every standard UI5 Tooling build task except for
`generateBundle`.
This PR aligns the behaviour also for the `generateBundle` task.

JIRA: CPOUI5FOUNDATION-724