[FIX] Resolve UI5 data directory relative to project
The UI5 data directory (previously called UI5 home directory) should
always be resolved relative to the project directory, which is where
the package.json file is located at.

This change also adds a list of options to the Configuration class to be
used by the CLI config command to allow all available config options to
be set or retrieved.

Follow-up of https://github.com/SAP/ui5-project/pull/635