[FIX] Ensure usage of provided UI5 data dir
Also moves check for framework libraries to earlier point as further
steps such as choosing a resolver and calling resolveVersion are only
needed when libraries are referenced.
This aligns the behavior with not having a framework configuration at
all but still passing "versionOverride" which is then also not
resolved/checked for correctness.