[FEATURE] TaskUtil: Add 'force' flag to cleanup task callback (#677)
JIRA: CPOUI5FOUNDATION-751
\@ui5/builder PR that depends on this change:
https://github.com/SAP/ui5-builder/pull/953

When cleaning up after tasks run, it can happen in couple cases:
- When the project is build
- When there's a SIGTERM signal
- (Maybe) on some error

We need to be able to distinguish between those cases as in some places
we rely on graceful termination and cleanup. Especially, after the
project has been build successfully. On the other hand when the user
presses `Ctrl + C`, for example, we need to enforce the termination.

This cahnge provides this context.

---------

Co-authored-by: Merlin Beutlberger <m.beutlberger@sap.com>
Co-authored-by: Günter Klatt <57760635+KlattG@users.noreply.github.com>