[BREAKING] Make '@ui5/builder' an optional peerDependency
This decoupling allows the use of @ui5/project without having to install
@ui5/builder with all its transitive dependencies in cases where no
build functionality is required. An example for that is UI5 linter.

BREAKING CHANGE:
Consumers of the Node.js API that make use of the ProjectGraph#build
method need to declare a dependency to '@ui5/builder' in their
respective package.json. The package manager should ensure that the
version fullfills the range specified in the package.json of
'@ui5/project'.